settings editor modal

+++ apps/cli/cli-core/src/trpc.ts
@@ -50,6 +50,35 @@ export const cliApiRouter = t.router({
     return Promise.all(res);
   }),
 
+  getHook: t.procedure
+    .input(z.object({ name: z.string() }))
+    .query(async ({ input }) => {
+      const { name } = input;
+
+      if (!fs.existsSync(path.join(HOOK_PATH, `${name}.json`))) {
+        console.log(`\u001b[31m[ERROR] ${name}.json does not exist`);
+      }
+
+      const body = await fsPromises.readFile(
+        path.join(HOOK_PATH, `${name}.json`),
+        "utf-8"
+      );
+
+      let config;
+      if (fs.existsSync(path.join(HOOK_PATH, `${name}.config.json`))) {
+        config = await fsPromises.readFile(
+          path.join(HOOK_PATH, `${name}.config.json`),
+          "utf-8"
+        );
+      }
+
+      return {
+        name,
+        body,
+        config: config ? JSON.parse(config) : undefined,
+      };
+    }),
+
   openFolder: t.procedure
     .input(z.object({ path: z.string() }))
     .mutation(async ({ input }) => {
@@ -84,11 +113,13 @@ export const cliApiRouter = t.router({
         );
         return fetchedResult;
       } catch (e) {
-        console.log("[ERROR] FAILED TO SEND");
+        console.log("\u001b[31m[ERROR] FAILED TO SEND");
         if ((e as any).code === "ECONNREFUSED") {
-          console.log("[ERROR] Connection refused. Is the server running?");
+          console.log(
+            "\u001b[31m[ERROR] Connection refused. Is the server running?"
+          );
         } else {
-          console.log("[ERROR] Unknown error", e);
+          console.log("\u001b[31m[ERROR] Unknown error", e);
         }
         throw new Error("Connection refused. Is the server running?");
       }
@@ -105,17 +136,75 @@ export const cliApiRouter = t.router({
     .mutation(async ({ input }) => {
       const { name, body, config } = input;
       console.log(`[INFO] Creating ${name}.json`);
-      const hookFile = await fsPromises.writeFile(
-        path.join(HOOK_PATH, `${name}.json`),
-        body
-      );
+
+      await fsPromises.writeFile(path.join(HOOK_PATH, `${name}.json`), body);
       if (config?.url || config?.query || config?.headers) {
         console.log(`[INFO] Config specified, creating ${name}.config.json`);
 
-        const hookConfigFile = await fsPromises.writeFile(
+        await fsPromises.writeFile(
           path.join(HOOK_PATH, `${name}.config.json`),
           JSON.stringify(config, null, 2)
         );
+      }
+    }),
+
+  updateHook: t.procedure
+    .input(
+      z.object({
+        name: z.string(),
+        body: z.string(),
+        config: configValidator.optional(),
+      })
+    )
+    .mutation(async ({ input }) => {
+      const { name, body, config } = input;
+      console.log(`[INFO] updating ${name}.json`);
+
+      const existingBody = await fsPromises.readFile(
+        path.join(HOOK_PATH, `${name}.json`),
+        "utf-8"
+      );
+
+      const parsedBody = JSON.parse(existingBody);
+
+      const updatedBody = {
+        ...parsedBody,
+        ...JSON.parse(body),
+      };
+
+      await fsPromises.writeFile(
+        path.join(HOOK_PATH, `${name}.json`),
+        JSON.stringify(updatedBody, null, 2)
+      );
+
+      if (config?.url || config?.query || config?.headers) {
+        console.log(`[INFO] Config specified, updating ${name}.config.json`);
+
+        // create config file if it doesn't exist
+        if (!fs.existsSync(path.join(HOOK_PATH, `${name}.config.json`))) {
+          await fsPromises.writeFile(
+            path.join(HOOK_PATH, `${name}.config.json`),
+            JSON.stringify(config, null, 2)
+          );
+        } else {
+          //update existing config file
+          const existingConfig = await fsPromises.readFile(
+            path.join(HOOK_PATH, `${name}.config.json`),
+            "utf-8"
+          );
+
+          const parsedConfig = JSON.parse(existingConfig);
+
+          const updatedConfig = {
+            ...parsedConfig,
+            ...config,
+          };
+
+          await fsPromises.writeFile(
+            path.join(HOOK_PATH, `${name}.config.json`),
+            JSON.stringify(updatedConfig, null, 2)
+          );
+        }
       }
     }),
 });
